Implemented search (only iOS 11+)
An empty search field will use the existent upcoming movies flow, while
a non-empty field will use the API for search. In both cases, the same
Driver triggers a next event that updates the data source, making the
UpcomingMoviesViewController effectively a more generic MoviesList (to
be renamed).

This commit is still missing some functionality to optimize the amount
of search requests while the user types.

Because of the way we accessed `self` from inside a closure in `init`,
we had to use `class` for the view model but now we can use `struct`.