Migrated TMDBModelTests to RxTest
- I prefer RxTest because I don't need to write expectations nor to
define a timeout for the expectation
- I don't like that RxTest doesn't seem to have a default easy way to
assert stuff in the events other than comparing the events themselves,
so if I want to know just the movies count (instead of all the movies),
I need to first map the events to the count. An extension was created to
more easily map Recorded<Event<T>> to Recorded<Event<U>>
- I don't like that RxTest doesn't seem to have a default easy way to
assert the presence of errors among the events. An extension was created
to simplify that.
- Using Swift 4.1 synthesized Equatable implementation for TMDBResults
and FailableMovie, whereas Movie implements its own definition because
comparing `id` is much simpler than comparing all of its fields.