Basic implementation of loading at view's bottom
- In the way it is implemented, it makes sense mostly for pagination,
while on search it doesn't feel very good to have a loading indicator at
the bottom.
- Added a failing unit test that should pass if Moya accepted a
TestScheduler parameter for the `delayedStub` behavior, however in the
way it is, it will try to delay actual 5 seconds instead of 5 virtual
time units. I noticed that Moya and Alamofire are too much for this
project's needs and at the same time they lacked power when integrating
with RxTest, so probably my next bet will be to write a network layer
from scratch.