Added rotation UI test
This should fix the "failing" build (rejected by Codecov because the
code coverage decreased in the most recent commit) as the code that was
added is now covered. Also adding iPad as part of the destinations
(the tests are run in parallel to iPhone) since this test doesn't make
much sense on iPhone, where the cells will be bigger on landscape thanks
to auto-layout, not to the layout invalidation. Using iOS 10 for iPad
given that iPhone is already using iOS 11 and the app supports older iOS
versions.