MockURLSessionDataTask now lives until completion
- It makes more sense now that the task is kept alive until it really
completes, and to do that, the task is now the responsible for telling
it has completed, instead of the MockURLSession.
- Made MockURLSessionDataTask internal so we can truly *unit* test it
(without being encapsulated by a Provider). Added unit tests that prove
the functionality of `resume()` and `cancel()`.
- Added unit tests for Rx resources to prove that this approach also
better handles memory.
- Also removed `isCancelled`: that logic shouldn't have been included.
It was used as a way to distinguish between paused and cancelled states,
but this is now done by calling `onCompleted()` on `isRunning` to signal
that no more events will be produced.