Fixed reentrancy anomaly
Replaced NSRecursiveLock by a serial DispatchQueue to ensure that all
MockURLSessionDataTask methods run in the same queue instead of multiple
thread trying to acquire a lock. This was leading RxSwift to detect a
reentrancy anomaly because the events called on `isRunning` inside
`complete()` were concurring with the one from `resume()` when using
delay=0.
Also removed `isRunning.onNext(false)` from `complete()` since it's
obvious that when it completes, it's not running anymore.