Replaced own DispatchQueue by SchedulerType's schedule
The unit tests were crashing because the DispatchQueue created for
MockURLSessionDataTask had qos=.background, so it would always execute
out of the main queue, while the TestScheduler requires in runtime that
its methods are running in the main queue. This approach here lets the
scheduler's queue be used, regardless of the chosen scheduler.