Fixed unit tests failing because of wrong expected time
The TestScheduler uses simulateProcessingDelay=true by default, which
means even immediate scheduling will increment 1 to the virtual time.
Because the MockURLSessionDataTask calls `schedule()`, each occurrence
of this would increment the time, making the original time expectations
wrong. This is fixed by setting simulateProcessingDelay to false in the
tests classes somehow involving the Provider.