ViewModel refactored to use functions
The previous structure was ok for a project this small, because the
ViewModel had a single output computed from a single set of inputs. In
reality though, ViewModels tend to be bigger, with more outputs
dependent on different sets of inputs. This new structure is better
suited for the real world.